requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
pikpakapi==0.1.11
python-dotenv==1.0.1
aiosqlite==0.20.0
